projects/ai-influencer/infrastructure/pipeline/requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
runpod==1.7.4
